requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
pypdf
python-dotenv